requirements.txt
pyglet
pillow
moderngl
numpy
pyrr
imgui
torch
scipy
plyfile
torchvision
opencv-python
tqdm
open3d
setuptools
transformers
diffusers
dearpygui
flask
comtypes
pyperclip
datasets
imgui[full]
moderngl-window
deprecated
trimesh
objloader
pytz
pvlib
imnodeeditor
yaml
pyyaml
blobfile
numba
matplotlib
pycuda